(PC-14097) fix(firebase): revert universal interface
