+++ jest/jest.setup.ts
@@ -69,8 +69,8 @@ jest.mock('libs/subcategories/mappings')
 jest.mock('libs/subcategories/useCategoryId')
 jest.mock('libs/subcategories/useSubcategory')
 
-/* See the corresponding mock in libs/firebase/app/__mocks__ */
-jest.mock('libs/firebase/app/index.web')
+/* See the corresponding mock in libs/firebase/__mocks__ */
+jest.mock('libs/firebase/firebase')
 
 /* See the corresponding mocks in libs/campaign/__mocks__ */
 jest.mock('libs/campaign')
