+++ package.json
@@ -103,7 +103,7 @@
     "babel-core": "^6.26.3",
     "color-alpha": "^1.1.3",
     "contentful-resolve-response": "^1.3.0",
-    "firebase": "^8.10.1",
+    "firebase": "^8.9.1",
     "geojson": "^0.5.0",
     "highlight-words-core": "^1.2.2",
     "intl": "^1.2.5",
