+++ src/features/identityCheck/components/FastEduconnectConnectionRequestModal.tsx
@@ -6,7 +6,7 @@ import { IdentityCheckMethod } from 'api/gen'
 import { useIdentityCheckContext } from 'features/identityCheck/context/IdentityCheckContextProvider'
 import { analytics } from 'libs/analytics'
 import { env } from 'libs/environment'
-import { useUbbleETAMessage } from 'libs/firebaseImpl/firestore/ubbleETAMessage'
+import { useUbbleETAMessage } from 'libs/firestore/ubbleETAMessage'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ButtonQuaternaryBlack } from 'ui/components/buttons/ButtonQuaternaryBlack'
 import { ButtonTertiaryBlack } from 'ui/components/buttons/ButtonTertiaryBlack'
