+++ src/features/maintenance/useMaintenance.tsx
@@ -1,10 +1,6 @@
 import { useEffect, useState } from 'react'
 
-import {
-  MAINTENANCE,
-  Maintenance,
-  maintenanceStatusListener,
-} from 'libs/firebaseImpl/firestore/maintenance'
+import { MAINTENANCE, Maintenance, maintenanceStatusListener } from 'libs/firestore/maintenance'
 
 export const useMaintenance = (): Maintenance => {
   const [maintenance, setMaintenance] = useState<Maintenance>({
