+++ src/libs/ABTesting/ABTesting.services.ts
@@ -1,4 +1,4 @@
-import firebaseRemoteConfig from 'libs/firebase/remote-config'
+import firebaseRemoteConfig from '@react-native-firebase/remote-config'
 
 import { DEFAULT_REMOTE_CONFIG } from './ABTesting.constants'
 import { CustomRemoteConfig, GenericRemoteConfig } from './ABTesting.types'
@@ -13,7 +13,7 @@ export const abTesting = {
   },
   /**
   `getValues()` does not fetch anything, it accesses the values fetched by `firebaseRemoteConfig().fetchAndActivate()`.
-  So be sure to have your remote config refreshed before using `getValues()`.
+  So be sure to have your remote config refreshed before using `getValues()`. 
   *
   * @returns object with remote config params.
   */
