--- src/libs/analytics/constants.ts
@@ -1,7 +1,0 @@
-import { AgentType } from 'api/gen'
-
-export const AGENT_TYPE = AgentType.agent_mobile
-
-export const EVENT_PAGE_VIEW_NAME = 'screen_view'
-
-export const EVENT_PAGE_VIEW_PARAM_KEY = 'screen_name'
