--- src/libs/firebase/app/index.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line no-restricted-imports
-import app from '@react-native-firebase/app'
-
-export const getFirebaseApp = () => app
