--- src/libs/firebase/dynamic-links/index.ts
@@ -1,4 +1,0 @@
-/* eslint-disable no-restricted-imports */
-export type { FirebaseDynamicLinksTypes } from '@react-native-firebase/dynamic-links'
-export { default } from '@react-native-firebase/dynamic-links'
-/* eslint-enable no-restricted-imports */
