--- src/libs/firebase/firestore/index.ts
@@ -1,4 +1,0 @@
-/* eslint-disable no-restricted-imports */
-export type { FirebaseFirestoreTypes } from '@react-native-firebase/firestore'
-export { default } from '@react-native-firebase/firestore'
-/* eslint-enable no-restricted-imports */
