--- src/libs/firebase/firestore/index.web.ts
@@ -1,3 +1,0 @@
-import { getFirebaseApp } from 'libs/firebase/app'
-
-export default () => getFirebaseApp().firestore()
