--- src/libs/firebase/remote-config/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line no-restricted-imports
-export { default } from '@react-native-firebase/remote-config'
