--- src/libs/firebaseImpl/firebaseConfig.ts
@@ -1,10 +1,0 @@
-import { env } from 'libs/environment'
-
-export const FIREBASE_CONFIG = {
-  apiKey: env.FIREBASE_APIKEY,
-  authDomain: env.FIREBASE_AUTHDOMAIN,
-  projectId: env.FIREBASE_PROJECTID,
-  storageBucket: env.FIREBASE_STORAGEBUCKET,
-  messagingSenderId: env.FIREBASE_MESSAGINGSENDERID,
-  appId: env.FIREBASE_APPID,
-}
