--- src/libs/firebaseImpl/firestore/client.ts
@@ -1,3 +1,0 @@
-import firestore from 'libs/firebase/firestore'
-
-export const firestoreRemoteStore = firestore()
