+++ src/libs/firestore/client.ts
@@ -1,0 +1,3 @@
+import firestore from '@react-native-firebase/firestore'
+
+export const firestoreRemoteStore = firestore()
