+++ src/libs/firestore/client.web.ts
@@ -1,0 +1,5 @@
+import { getFirebaseApp } from 'libs/firebase'
+import 'firebase/firestore'
+
+const firebaseApp = getFirebaseApp()
+export const firestoreRemoteStore = firebaseApp.firestore()
