+++ src/libs/firestore/maintenance.ts
@@ -1,8 +1,8 @@
 import { t } from '@lingui/macro'
 
 import { env } from 'libs/environment'
-import { firestoreRemoteStore } from 'libs/firebaseImpl/firestore/client'
-import { RemoteStoreCollections, RemoteStoreDocuments } from 'libs/firebaseImpl/firestore/types'
+import { firestoreRemoteStore } from 'libs/firestore/client'
+import { RemoteStoreCollections, RemoteStoreDocuments } from 'libs/firestore/types'
 import { captureMonitoringError } from 'libs/monitoring'
 
 export enum MAINTENANCE {
